HeaderButtonMetrics refactor
