+++ modules/javafx.graphics/src/main/java/com/sun/javafx/stage/HeaderButtonMetrics.java
@@ -23,9 +23,10 @@
  * questions.
  */
 
-package com.sun.glass.ui;
+package com.sun.javafx.stage;
 
 import javafx.geometry.Dimension2D;
+import javafx.scene.layout.HeaderBar;
 import javafx.stage.StageStyle;
 import java.util.Objects;
 
@@ -35,11 +36,9 @@ import java.util.Objects;
  * @param leftInset the size of the left inset
  * @param rightInset the size of the right inset
  * @param minHeight the minimum height of the window buttons
- * @see HeaderButtonOverlay
+ * @see HeaderBar
  */
 public record HeaderButtonMetrics(Dimension2D leftInset, Dimension2D rightInset, double minHeight) {
-
-    public static HeaderButtonMetrics EMPTY = new HeaderButtonMetrics(new Dimension2D(0, 0), new Dimension2D(0, 0), 0);
 
     public HeaderButtonMetrics {
         Objects.requireNonNull(leftInset);
@@ -48,13 +47,5 @@ public record HeaderButtonMetrics(Dimension2D leftInset, Dimension2D rightInset,
         if (minHeight < 0) {
             throw new IllegalArgumentException("minHeight cannot be negative");
         }
-    }
-
-    public double totalInsetWidth() {
-        return leftInset.getWidth() + rightInset.getWidth();
-    }
-
-    public double maxInsetHeight() {
-        return Math.max(leftInset.getHeight(), rightInset.getHeight());
     }
 }
